app/frontend/requirements.txt
streamlit>=1.24.0
requests>=2.28.1
pandas>=1.5.0
plotly>=5.10.0
pillow>=9.2.0 